20260403: Fix business date fields (int64→string), fix money fields (float64→int64), add line/location_area Postgres adapters, fix subscription plan JOIN for client detail, update entity IDs

+++ contrib/postgres/internal/adapter/fulfillment/work_item.go
@@ -333,7 +333,7 @@ func (r *PostgresFulfillmentRepository) GetFulfillmentListPageData(
 			status            string
 			providerStatus    string
 			providerReference string
-			deliveryCost      float64
+			deliveryCost      int64
 			currency          string
 			expenditureID     sql.NullString
 			scheduledAt       sql.NullTime
@@ -489,7 +489,7 @@ func (r *PostgresFulfillmentRepository) GetFulfillmentItemPageData(
 		status            string
 		providerStatus    string
 		providerReference string
-		deliveryCost      float64
+		deliveryCost      int64
 		currency          string
 		expenditureID     sql.NullString
 		supplierName      string
